Move things around (properties and traits)
