+++ properties/traits.py
@@ -5,156 +5,8 @@ from __future__ import unicode_literals
 
 import traitlets as tr
 from six import with_metaclass
-from six import integer_types
 from six import iteritems
-import numpy as np
-
-
-class Property(object):
-    """
-        Base property class that establishes property behavior
-    """
-
-    info_text = 'corrected'
-    name = ''
-    class_name = ''
-
-    def __init__(self, help, **kwargs):
-        self._base_help = help
-        for key in kwargs:
-            if key[0] == '_':
-                raise AttributeError(
-                    'Cannot set private property: "{}".'.format(key))
-            if not hasattr(self, key):
-                raise AttributeError(
-                    'Unknown key for property: "{}".'.format(key))
-            setattr(self, key, kwargs[key])
-
-    @property
-    def help(self):
-        if getattr(self, '_help', None) is None:
-            self._help = self._base_help
-        return self._help
-
-    def info(self):
-        return self.info_text
-
-    @property
-    def default(self):
-        """default value of the property"""
-        return getattr(self, '_default', None)
-
-    @default.setter
-    def default(self, value):
-        self._default = value
-
-    @property
-    def required(self):
-        """required properties must be set for validation to pass"""
-        return getattr(self, '_required', False)
-
-    @required.setter
-    def required(self, value):
-        assert isinstance(value, bool), "Required must be a boolean."
-        self._required = value
-
-    def is_valid(self, scope):
-        attr = getattr(scope, self.name, None)
-        if (attr is None) and self.required:
-            raise ValueError(self.name)
-
-    def validate(self, instance, value):
-        """validates the property attributes"""
-        return value
-
-    def get_property(self):
-        """establishes access of property values"""
-
-        scope = self
-
-        def fget(self):
-            return self._get(scope.name)
-
-        def fset(self, value):
-            value = scope.validate(self, value)
-            self._set(scope.name, value)
-            # self._on_property_change(
-            #     dict(
-            #         name=scope.name,
-            #         value=value
-            #     )
-            # )
-
-        return property(fget=fget, fset=fset, doc=scope.help)
-
-    def as_json(self, value):
-        return value
-
-    def from_json(self, value):
-        return value
-
-    def error(self, instance, value):
-        raise ValueError(
-            "The '{name}' trait of a {cls} instance must be {info}. "
-            "A value of {val!r} {vtype!r} was specified.".format(
-                name=self.name,
-                cls=instance.__class__.__name__,
-                info=self.info(),
-                val=value,
-                vtype=type(value)
-            )
-        )
-
-    def get_backend(self, backend):
-        self.new_backend()  # Ensures that the backends has been instantiated.
-        if backend not in self._backends:
-            raise Exception(
-                'The "{backend}" backend is not supported '
-                'for a {cls} property.'.format(
-                    backend=backend,
-                    cls=self.__class__.__name__
-                )
-            )
-        if self._backends[backend] is None:
-            return None
-        return self._backends[backend](self)
-
-    @classmethod
-    def new_backend(cls, backend=None):
-        if getattr(cls, '_backends', None) is None:
-            cls._backends = {
-                "dict": None
-            }
-        if backend is None:
-            return
-
-        def new_backend(func):
-            # print('adding {} backend'.format(backend))
-            cls._backends[backend] = func
-        return new_backend
-
-
-class Integer(Property):
-
-    info_text = 'an integer'
-
-    def validate(self, instance, value):
-        if isinstance(value, float) and np.isclose(value, int(value)):
-            value = int(value)
-        if not isinstance(value, integer_types):
-            self.error(instance, value)
-        return int(value)
-
-    def as_json(self, value):
-        if value is None or np.isnan(value):
-            return None
-        return int(np.round(value))
-
-    def from_json(self, value):
-        return int(str(value))
-
-    def get_traitlets_backend(self):
-        return tr.Int()
+from . import properties
 
 
 def resolve(attr, bases, classdict):
@@ -193,7 +45,7 @@ class PropertyMetaclass(type):
         prop_dict = {
             key: value for key, value in classdict.items()
             if (
-                isinstance(value, Property)
+                isinstance(value, properties.Property)
             )
         }
 
@@ -343,12 +195,3 @@ def HasProperties(backend=None):
 
 #         return property(fget=fget, fset=fset, doc=self.help)
 
-
-
-
-import traitlets as tr
-
-
-@Integer.new_backend('traitlets')
-def get_int(prop):
-    return tr.Integer(help=prop.help)
